web app added

+++ Alzheimers Detection/Web-App/app.py
@@ -1,11 +1,21 @@
 import streamlit as st
 from PIL import Image
-import pandas as pd
-import matplotlib.pyplot as plt
-import requests
-import json
+import numpy as np
 import os
 import tempfile
+import tensorflow as tf
+
+# Function to load the model
+@st.cache_resource
+def load_model():
+    script_dir = os.path.dirname(__file__)  
+    rel_path = 'model.h5'  
+    abs_model_path = os.path.join(script_dir, rel_path)  
+    model = tf.keras.models.load_model(abs_model_path)
+    return model
+
+
+model = load_model()
 
 st.title("Alzheimer's Detection")
 
@@ -18,30 +28,42 @@ The disease is classified into different stages based on the severity of symptom
 Non Demented, and Very Mild Demented.
 """)
 
-# Handle the path to the sample image using os library
+st.header("Sample Image")
 script_dir = os.path.dirname(__file__)
 rel_path = "../Images/image.jpg"
 abs_file_path = os.path.join(script_dir, rel_path)
-st.header("Sample Image")
 try:
     image = Image.open(abs_file_path)
     st.image(image, caption='Sample Brain Scan Image')
 except Exception as e:
     st.write(f"Error loading image: {e}")
-# Handle the path to the uploaded image using os library
+
 st.header("Upload Your MRI Scan")
 scan = st.file_uploader("Choose an image...", type=["jpg", "jpeg", "png"])
+
+def preprocess_image(image, target_size):
+    image = image.convert("RGB")
+    image = image.resize(target_size)
+    image_array = tf.keras.preprocessing.image.img_to_array(image)
+    image_array = np.expand_dims(image_array, axis=0)
+    return image_array
+
 if scan is not None:
-    try:
-        
-        with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
-            tmp_file.write(scan.read())
-            tmp_path = tmp_file.name
+    if st.button('Upload'):
+        try:
+            with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
+                tmp_file.write(scan.read())
+                tmp_path = tmp_file.name
 
-        img = Image.open(tmp_path)
-        st.image(img, caption='Uploaded Brain Scan Image')
-        
-        os.remove(tmp_path)
-    except Exception as e:
-        st.write(f"Error processing uploaded image: {e}")
+            img = Image.open(tmp_path)
+            st.image(img, caption='Uploaded Brain Scan Image')
+            img_array = preprocess_image(img, target_size=(176, 176)) 
+            predictions = model.predict(img_array)
+            predicted_class = np.argmax(predictions, axis=1)
+            class_labels = ['MildDemented', 'ModerateDemented', 'NonDemented', 'VeryMildDemented']
+            result = class_labels[predicted_class[0]]
+            st.markdown(f'<h3 style="color: red;">Prediction: {result}</h3>', unsafe_allow_html=True)
 
+            os.remove(tmp_path)
+        except Exception as e:
+            st.write(f"Error processing uploaded image: {e}")
